Create /pages directory
