+++ web/src/pages/ConfirmationPage.js
@@ -1,8 +1,8 @@
 import React from 'react'
 import { css } from 'emotion'
-import { theme, H1, H2 } from './styles'
+import { theme, H1, H2 } from '../styles'
 import { Trans } from 'lingui-react'
-import Layout from './components/Layout'
+import Layout from '../components/Layout'
 
 const contentClass = css`
   h1 {
