+++ web/src/pages/FourOhFourPage.js
@@ -2,8 +2,8 @@ import React from 'react'
 import { Trans } from 'lingui-react'
 import { css } from 'react-emotion'
 import { NavLink } from 'react-router-dom'
-import { H1, theme } from './styles'
-import Layout from './components/Layout'
+import { H1, H2, theme } from '../styles'
+import Layout from '../components/Layout'
 
 const contentClass = css`
   p {
@@ -11,20 +11,24 @@ const contentClass = css`
   }
 `
 
-class ErrorPage extends React.Component {
+class FourOhFourPage extends React.Component {
   render() {
     return (
       <Layout contentClass={contentClass}>
         <H1>
-          <Trans>Request failed</Trans>
+          <Trans>404: Page not found</Trans>
         </H1>
+        <H2>
+          <Trans>
+            Oops! We can’t seem to find the page you are looking for:
+          </Trans>
+        </H2>
         <p>
           <Trans>
-            Please contact IRCC directly in order to change your appointment.
+            If you wish to restart the rescheduling process, please navigate to
+            the Home Page and start the process again.
           </Trans>
         </p>
-        <p>vancouverIRCC@cic.gc.ca</p>
-        <p>1-888-242-2100</p>
         <NavLink to="/">
           <Trans>← Go Back</Trans>
         </NavLink>
@@ -33,4 +37,4 @@ class ErrorPage extends React.Component {
   }
 }
 
-export default ErrorPage
+export default FourOhFourPage
