+++ web/src/pages/LandingPage.js
@@ -1,9 +1,9 @@
 import React from 'react'
 import styled, { css } from 'react-emotion'
 import { NavLink } from 'react-router-dom'
-import { H1, H2, H3, theme, mediaQuery } from './styles'
-import Layout from './components/Layout'
-import { buttonStyles } from './components/forms/Button'
+import { H1, H2, H3, theme, mediaQuery } from '../styles'
+import Layout from '../components/Layout'
+import { buttonStyles } from '../components/forms/Button'
 import { Trans } from 'lingui-react'
 import { Helmet } from 'react-helmet'
