+++ web/src/pages/ReviewPage.js
@@ -3,13 +3,13 @@ import { css } from 'react-emotion'
 import { Trans } from 'lingui-react'
 import { Query } from 'react-apollo'
 import { NavLink, Redirect } from 'react-router-dom'
-import { H1, theme, BottomContainer, TopContainer } from './styles'
-import Layout from './components/Layout'
-import { GET_USER_DATA, SUBMIT } from './queries'
-import Button from './components/forms/Button'
-import Summary from './components/Summary'
-import { Submission } from './components/Submission'
-import { Reminder } from './components/Reminder'
+import { H1, theme, BottomContainer, TopContainer } from '../styles'
+import Layout from '../components/Layout'
+import { GET_USER_DATA, SUBMIT } from '../queries'
+import Button from '../components/forms/Button'
+import Summary from '../components/Summary'
+import { Submission } from '../components/Submission'
+import { Reminder } from '../components/Reminder'
 
 const contentClass = css`
   p {
