+++ web/src/pages/__tests__/RegistrationPage.test.js
