+++ web/src/routes.js
@@ -7,7 +7,7 @@ export default [
     path: '/',
     exact: true,
     component: asyncComponent({
-      loader: () => import('./LandingPage'),
+      loader: () => import('./pages/LandingPage'),
       // eslint-disable-next-line react/display-name
       Placeholder: () => <div>...LOADING...</div>, // this is optional, just returns null by default
     }),
@@ -16,42 +16,42 @@ export default [
     path: '/register',
     exact: true,
     component: asyncComponent({
-      loader: () => import('./RegistrationPage'),
+      loader: () => import('./pages/RegistrationPage'),
     }),
   },
   {
     path: '/calendar',
     exact: true,
     component: asyncComponent({
-      loader: () => import('./CalendarPage'),
+      loader: () => import('./pages/CalendarPage'),
     }),
   },
   {
     path: '/review',
     exact: true,
     component: asyncComponent({
-      loader: () => import('./ReviewPage'),
+      loader: () => import('./pages/ReviewPage'),
     }),
   },
   {
     path: '/confirmation',
     exact: true,
     component: asyncComponent({
-      loader: () => import('./ConfirmationPage'),
+      loader: () => import('./pages/ConfirmationPage'),
     }),
   },
   {
     path: '/error',
     exact: true,
     component: asyncComponent({
-      loader: () => import('./ErrorPage'),
+      loader: () => import('./pages/ErrorPage'),
     }),
   },
   {
     path: '*',
     exact: true,
     component: asyncComponent({
-      loader: () => import('./FourOhFourPage'),
+      loader: () => import('./pages/FourOhFourPage'),
     }),
   },
 ]
